pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "selenium-ui-test-tool"
version = "1.0.0"
description = "Bibliothèque Python pour faciliter les tests UI automatisés avec Selenium"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Yann Dipita", email = "dipitay@gmail.com"}
]
keywords = ["selenium", "testing", "ui", "automation", "webdriver"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "selenium>=4.15.0",
    "python-dotenv>=1.0.0",
    "webdriver-manager>=4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools]
packages = {find = {}}

